+++ src/DefaultConstants.php
@@ -14,4 +14,10 @@ abstract class DefaultConstants
      * @var float
      */
     const π = M_PI;
+
+    /**
+     *
+     * @var float
+     */
+    const e = M_E;
 }