+++ src/Plane/Curves/EllipseInterface.php
@@ -8,8 +8,39 @@ namespace samizdam\Geometry\Plane\Curves;
  */
 interface EllipseInterface extends CircleInterface
 {
-    
-    public function getMinorAxis();
 
-    public function getMajorAxis();
+    /**
+     * Get Semi-Minor Axis value
+     * 
+     * @return float
+     */
+    public function getSemiMajorAxis();
+
+    /**
+     * Get Semi-Major Axis value
+     * 
+     * @return float
+     */
+    public function getSemiMinorAxis();
+
+    /**
+     * Get Focal Paramert value
+     *
+     * @return float
+     */
+    public function getFocalParameter();
+    
+    /**
+     * Get Eccentricity value
+     * 
+     * @return float 
+     */
+    public function getEccentricity();
+    
+    /**
+     * Get Focal Length value
+     * 
+     * @return float
+     */
+    public function getFocalLength();
 }